+++ sql/postgresql/initialization/V1.4.0_PG.sql
@@ -9365,7 +9365,7 @@ COMMENT ON TABLE public.mc_task_scope IS 'mc_task_scope';
 DROP TABLE IF EXISTS public.message;
 
 CREATE TABLE public.message (
-  id integer NOT NULL,
+  id BIGINT NOT NULL,
   sender_id BIGINT,
   receiver_id BIGINT,
   title varchar NOT NULL,
